Set all int features to use int64
as 'nn.Embedding' cannot accept bool, uint8, int16.
int64 is used as the old code used that.

+++ geognn/features/base_feature_classes.py
@@ -97,7 +97,7 @@ class LabelEncodedFeature(Feature):
         name: str,
         possible_values: list[Any],
         get_raw_value: Callable[[Atom, Mol, Conformer, DGLGraph | None], Any],
-        dtype: torch.dtype = torch.uint8,
+        dtype: torch.dtype = torch.int64,
     ) -> 'LabelEncodedFeature':
         return LabelEncodedFeature(
             name = name,
@@ -115,7 +115,7 @@ class LabelEncodedFeature(Feature):
         name: str,
         possible_values: list[Any],
         get_raw_value: Callable[[Bond, Mol, Conformer, DGLGraph | None], Any],
-        dtype: torch.dtype = torch.uint8,
+        dtype: torch.dtype = torch.int64,
     ) -> 'LabelEncodedFeature':
         return LabelEncodedFeature(
             name = name,
